hotel_booking_system/hotel_comments_service/requirements.txt
google-cloud-firestore
fastapi
uvicorn
python-dotenv